+++ R/write_codemeta.R
@@ -18,7 +18,8 @@
 #' @param force_update Update guessed fields even if they are defined in an
 #'   existing codemeta.json file
 #' @param use_git_hook Whether to create a pre-commit hook requiring
-#'   codemeta.json to be updated when DESCRIPTION is changed.  Defaults to TRUE.
+#'   codemeta.json to be updated when DESCRIPTION is changed. Default is \code{FALSE} to avoid
+#'   an unwanted alteration of the user's git environment.
 #' @param verbose Whether to print messages indicating opinions e.g. when
 #'   DESCRIPTION has no URL. See \code{\link{give_opinions}}.
 #' @param ...  additional arguments to \code{\link{write_json}}
@@ -41,7 +42,7 @@
 #' }
 write_codemeta <- function(
   pkg = ".", path = "codemeta.json", root = ".", id = NULL, use_filesize = FALSE,
-  force_update = getOption("codemeta_force_update", TRUE), use_git_hook = TRUE,
+  force_update = getOption("codemeta_force_update", TRUE), use_git_hook = FALSE,
   verbose = TRUE, ...
 ) {